[cleanup ]cleanup imports and restructure files (#9)
removed ununsed imports, restructured codebase
