+++ package.json
@@ -13,7 +13,6 @@
     "ts:check": "tsc"
   },
   "dependencies": {
-    "@react-native-async-storage/async-storage": "^1.19.3",
     "@react-navigation/native": "^6.1.8",
     "@react-navigation/native-stack": "^6.9.14",
     "@supabase/supabase-js": "^2.38.0",
